+++ src/components/Molecules/Promo/Promo.style.js
@@ -1,5 +1,5 @@
 import styled, { css } from 'styled-components';
-import container from '../../../theme/shared/containers';
+import container from '../../../theme/shared/allContainers';
 import spacing from '../../../theme/shared/spacing';
 import zIndex from '../../../theme/shared/zIndex';
